+++ kardscm/web/__init__.py
@@ -1,0 +1,1 @@
+"""Local webUI for kardscm collection management."""